Reworked height parameter
Previous height_difference_score calculation turned out to be pretty bad,
the agent played better when it's value was 0. So I changed it entirely.
It's now a sum of deltas, obtained from comparing each value of floor to it's average.
Turns out that works pretty good, and the agent consistently reaches over 200 points
before desyncing.