Improves decision time with placement cache (WIP)
When analysing pieces for LOOK_AHEAD, we calculate all their possible
placements, and then their scores. We then would discard this and recalculate them
when said piece arrived. This commit makes it so that the calculated placements
for the future piece are now stored, so that they may be retrieved when it arrives as
possibilities for placement should not change. This cuts decision time a fair amount.
The effect should be much more noticeable with higher LOOK_AHEAD values, once that
is implemented (that will require some changes to the way placements are stored and retrieved)

Score is still recalculated, to take future pieces into account.

This does not seem to be working properly, as the agent seems to be playing much worse.
Maybe due to some oversight this is not working as intended. Still, it's a start.